Updated EmailHeaderExtractor script source
This update brings along, a new domainExtraction function that is more efficient now than the previous version. And a little adjustment to the output of the data, where the List of IPs has been named Email Traceroute to more accurately represent what it outputs, changed the way the origin_ip variable gets used, as it use to output None when in actual fact the last IP on the traceroute is usually the source. Also added an exception incase ip-api.com can't handle the API request

Bug Fixes:
- Fixed a bug with the domainExtraction function